+++ StopRugRemoval/StopRugRemoval/HarmonyPatches/BombHandling/ConfirmBomb.cs
@@ -1,19 +1,32 @@
 ﻿using System.Reflection;
 using AtraBase.Toolkit.Reflection;
+using AtraShared.Utils;
 using HarmonyLib;
 using Microsoft.Xna.Framework;
+using StardewModdingAPI.Utilities;
 
 namespace StopRugRemoval.HarmonyPatches.BombHandling;
 
+/// <summary>
+/// Class to hold patches against answerDialogueAction to handle bomb confirmation.
+/// </summary>
 [HarmonyPatch]
 internal static class ConfirmBomb
 {
+    internal static readonly PerScreen<bool> HaveConfirmed = new(createNewState: () => false);
+    internal static readonly PerScreen<Vector2> BombLocation = new(createNewState: () => Vector2.Zero);
+    internal static readonly PerScreen<int> WhichBomb = new(createNewState: () => 0);
 
+    /// <summary>
+    /// Defines the methods for which to patch.
+    /// </summary>
+    /// <returns>Methods to patch.</returns>
+    [UsedImplicitly]
     internal static IEnumerable<MethodBase> TargetMethods()
     {
         foreach (Type t in typeof(GameLocation).GetAssignableTypes(publiconly: true, includeAbstract: false))
         {
-            if (AccessTools.Method(t, nameof(GameLocation.answerDialogueAction), new Type[] { typeof(string), typeof(string[]) }) is MethodBase method
+            if (t.DeclaredInstanceMethodNamedOrNull(nameof(GameLocation.answerDialogueAction), new Type[] { typeof(string), typeof(string[]) }) is MethodBase method
                 && method.DeclaringType == t)
             {
                 yield return method;
@@ -21,122 +34,40 @@ internal static class ConfirmBomb
         }
     }
 
-    internal static bool Prefix(GameLocation __instance, string __0, string[] __1)
+    /// <summary>
+    /// Prefixes answerDialogueAction to handle the bomb questions.
+    /// </summary>
+    /// <param name="__instance">Gamelocation.</param>
+    /// <param name="__0">question_answer.</param>
+    /// <param name="__result">Result to substitute in for original function.</param>
+    /// <returns>True to continue to vanilla function, false to skip.</returns>
+    [UsedImplicitly]
+    [SuppressMessage("StyleCop", "SA1313", Justification = "Style prefered by Harmony")]
+    internal static bool Prefix(GameLocation __instance, string __0, ref bool __result)
     {
         try
         {
             switch (__0)
             {
                 case "atravitaInteractionTweaksBombs_BombsArea":
-                    SObjectPatches.HaveConfirmed.Value = true;
+                    HaveConfirmed.Value = true;
                     goto case "atravitaInteractionTweaksBombs_BombsYes";
                 case "atravitaInteractionTweaksBombs_BombsYes":
                     Game1.player.reduceActiveItemByOne();
-                    ExplodeBomb(__instance, SObjectPatches.whichBomb.Value, SObjectPatches.BombLocation.Value);
+                    GameLocationUtils.ExplodeBomb(__instance, WhichBomb.Value, BombLocation.Value, ModEntry.Multiplayer);
                     break;
                 case "atravitaInteractionTweaksBombs_BombsNo":
                     break;
                 default:
                     return true;
             }
+            __result = true;
+            return false;
         }
         catch (Exception ex)
         {
             ModEntry.ModMonitor.Log($"Ran into issues in prefix for confirming bombs.\n\n{ex}", LogLevel.Error);
         }
         return true;
-    }
-
-    /// <summary>
-    /// The code in this function is effectively copied from the game, and explodes a bomb on this tile.
-    /// </summary>
-    /// <param name="loc">Location to explode bomb.</param>
-    /// <param name="whichBomb">Which bomb to explode.</param>
-    /// <param name="tileloc">Tile to explode bomb on.</param>
-    internal static void ExplodeBomb(GameLocation loc, int whichBomb, Vector2 tileloc)
-    {
-        int bombID = Game1.random.Next();
-        loc.playSound("thudStep");
-        TemporaryAnimatedSprite tas_bomb = new(
-            initialParentTileIndex: whichBomb,
-            animationInterval: 100f,
-            animationLength: 1,
-            numberOfLoops: 24,
-            position: tileloc,
-            flicker: true,
-            flipped: false,
-            parent: loc,
-            owner: Game1.player)
-        {
-            shakeIntensity = 0.5f,
-            shakeIntensityChange = 0.002f,
-            extraInfoForEndBehavior = bombID,
-            endFunction = loc.removeTemporarySpritesWithID,
-        };
-        Multiplayer mp = ModEntry.MultiPlayer;
-        mp.broadcastSprites(loc, tas_bomb);
-        TemporaryAnimatedSprite tas_yellow = new(
-            textureName: "LooseSprites\\Cursors",
-            sourceRect: new Rectangle(598, 1279, 3, 4),
-            animationInterval: 53f,
-            animationLength: 5,
-            numberOfLoops: 9,
-            position: tileloc + (new Vector2(5f, 3f) * 4f),
-            flicker: true,
-            flipped: false,
-            layerDepth: (float)(tileloc.Y + 7) / 10000f,
-            alphaFade: 0f,
-            color: Color.Yellow,
-            scale: 4f,
-            scaleChange: 0f,
-            rotation: 0f,
-            rotationChange: 0f)
-        {
-            id = bombID,
-        };
-        mp.broadcastSprites(loc, tas_yellow);
-        TemporaryAnimatedSprite tas_orange = new(
-            textureName: "LooseSprites\\Cursors",
-            sourceRect: new Rectangle(598, 1279, 3, 4),
-            animationInterval: 53f,
-            animationLength: 5,
-            numberOfLoops: 9,
-            position: tileloc + (new Vector2(5f, 3f) * 4f),
-            flicker: true,
-            flipped: false,
-            layerDepth: (float)(tileloc.Y + 7) / 10000f,
-            alphaFade: 0f,
-            color: Color.Orange,
-            scale: 4f,
-            scaleChange: 0f,
-            rotation: 0f,
-            rotationChange: 0f)
-        {
-            delayBeforeAnimationStart = 100,
-            id = bombID,
-        };
-        mp.broadcastSprites(loc, tas_orange);
-        TemporaryAnimatedSprite tas_white = new(
-            textureName: "LooseSprites\\Cursors",
-            sourceRect: new Rectangle(598, 1279, 3, 4),
-            animationInterval: 53f,
-            animationLength: 5,
-            numberOfLoops: 9,
-            position: tileloc + (new Vector2(5f, 3f) * 4f),
-            flicker: true,
-            flipped: false,
-            layerDepth: (float)(tileloc.Y + 7) / 10000f,
-            alphaFade: 0f,
-            color: Color.White,
-            scale: 4f,
-            scaleChange: 0f,
-            rotation: 0f,
-            rotationChange: 0f)
-        {
-            delayBeforeAnimationStart = 200,
-            id = bombID,
-        };
-        ModEntry.MultiPlayer.broadcastSprites(loc, tas_white);
-        loc.netAudio.StartPlaying("fuse");
     }
 }