+++ StopRugRemoval/StopRugRemoval/ModConfig.cs
@@ -22,12 +22,12 @@ public class ModConfig
     /// Gets or sets a value indicating whether whether or not I should be able to place rugs under things.
     /// </summary>
     public bool CanPlaceRugsUnder { get; set; } = true;
+#endif
 
     /// <summary>
     /// Gets or sets a value indicating whether or not bombs should be confirmed.
     /// </summary>
     public bool ConfirmBombs { get; set; } = true;
-#endif
 
     /// <summary>
     /// Gets or sets a value indicating whether whether or not to prevent the removal of items from a table.
@@ -43,6 +43,11 @@ public class ModConfig
     /// Gets or sets a value indicating whether grass should be placed under objects.
     /// </summary>
     public bool PlaceGrassUnder { get; set; } = true;
+
+    /// <summary>
+    /// Gets or sets a value indicating whether jukeboxes should be playable everywhere.
+    /// </summary>
+    public bool JukeboxesEverywhere { get; set; } = true;
 
     /// <summary>
     /// Gets or sets keybind to use to remove an item from a table.